backend/requirements.txt
flask
python-pptx
requests
yake
wikipedia
